chore: move util func to utils file
