+++ ui/tooltip/src/TooltipContent.tsx
@@ -3,6 +3,7 @@ import * as TooltipPrimitive from "@radix-ui/react-tooltip";
 import { styled, keyframes, theme } from "@washingtonpost/wpds-theme";
 import type * as WPDS from "@washingtonpost/wpds-theme";
 import { TooltipContentProps as RadixTooltipContentProps } from "@radix-ui/react-tooltip";
+import { getPixelsFromRem } from "./utils";
 
 export const enum TOOLTIP_CONTENT_SIDE {
   left = "left",
@@ -46,36 +47,6 @@ type WPDSThemeSpaceObject = {
   value: string;
   scale: string;
   prefix: string;
-};
-
-// Radix offset only accepts a number of pixels it should be offsetting
-const getPixelsFromRem = (
-  valToConvert: number | string | WPDSThemeSpaceObject
-): number => {
-  const convertRemToPixels = (val) => (val as unknown as number) * 16;
-  // if we pass the number 0, return the number 0;
-  if (typeof valToConvert === "number" && valToConvert === 0) {
-    return 0;
-  }
-
-  // if we pass a string that includes the words rem
-  //mainly used for docs site examples and storybook example
-  if (typeof valToConvert === "string" && valToConvert.includes("rem")) {
-    const val = valToConvert.split("rem")[0];
-
-    return convertRemToPixels(val);
-  }
-
-  if (typeof valToConvert === "object") {
-    // if we pass in an object and it has a value
-    const val = valToConvert.value.split("rem")[0];
-
-    return convertRemToPixels(val);
-  }
-
-  throw Error(
-    "Please only provide 1) the number zero 2) a WPDS theme space object or 3) a string that follows the patter `xxrem` where `xx` is a number"
-  );
 };
 
 const StyledContentWrapper = styled(TooltipPrimitive.Content, {
